requirement.txt
yfinance
psycopg2
pandas
python-dotenv